requirements.txt
openai
pymongo
python-dotenv
langchain-openai
langchain-community

